requirements.txt
APScheduler==3.7.0
bidict==0.21.2
click==7.1.2
dnspython==1.16.0
eventlet==0.30.2
Flask==1.1.2
Flask-Cors==3.0.10
flask-serial==1.1.0
Flask-SocketIO==4.3.1
Flask-SQLAlchemy==2.5.1
greenlet==1.1.0
gunicorn==20.1.0
importlib-metadata==4.0.1
itsdangerous==1.1.0
Jinja2==2.11.3
MarkupSafe==1.1.1
numpy==1.20.3
picamera==1.13
pyserial==3.5
python-engineio==3.13.2
python-socketio==4.6.0
pytz==2021.1
six==1.16.0
SQLAlchemy==1.4.14
typing==3.7.4.3
typing-extensions==3.10.0.0
tzlocal==2.1
Werkzeug==1.0.1
zipp==3.4.1
